Allow sliders to exist outside a balance group
Not all sliders belong to a group which must sum to 100%. Interface
groups which consist of a single "type: slider" element will not be part
of a group balancer and may be edited by a user.

Alters the ChangeTrigger to read the value of the input associated with
a slider, rather than the ".slider" element which doesn't have a value.
Previously the "Save changes" form would be displayed by accident (even
though it is intended to show the form) by the balancer triggering a
change on another slider in the group, rather than by the slider the
user edited.