Fix poor performance when loading datasets
Fixes that loading the dataset interface after already having loaded a
dataset, would hang for tens of seconds when setting the new content
with $.fn.html.

It's not clear why jQuery - or the browser - hangs, but the problem was
present in both Chrome 69 and Firefox 62 (but significantly worse in
Chrome).

Closes #142