Round all group sliders when user makes change
When changing a slider belonging to a group which has not previously
been modified, those that are changed were rounded while those that were
not still retained many decimal placed. This results in groups not
summing to 100%.

Until a better balancing solution can be created (like ETModel), now all
sliders belonging to the group will be rounded, ensuring the group sums
to 100% at the cost of modifying inputs which the user may not have
explicitly touched.

Closes #155