Fix links to GitHub files which are symlinks
Instead of resolving GitFiles to their real path, in an attempt to show
the real file used by the dataset, instead continue to show the relative
path from the dataset or parent dataset, but _link_ to the real file on
GitHub.

This means when viewing a file whose parent dataset symlinks a file from
another dataset, we'll link to the correct file. For example:

  * Northern Ireland has no curve
  * UK symlinks a curve from NL

Previously we'd link to the non-existant file in the UK dataset. Now
we'll still show datasets/curves/uk/something.csv in the UI, but will
link to the curve in the NL dataset on GitHub.

This reverts commit 44632c40d38dba75bf9e07f381fa1eb0766306b2.
This reverts commit 661f42a780f4e6e679d0434a2cd78e8eb830d9e0.

Ref #314