Add support for ENTSO-based energy balances
Datasets which are based on an ENTSO energy balance will have their
`data_source` set to "entso". Any interface element with an `entso` key
will execute the value as a GQL query to read from the CSV for those
datasets with an "entso" `data_source`.

Currently the supported GQL is extremely limited: `EB` only.